fix: three independent test failures on main
- Delete app/cli/cli-segment-writes.test.ts: a leftover duplicate from the
  Segment→Beat CLI rename (#1156) that never got deleted. It still called
  the removed `segment` CLI noun, which now prints the root help instead
  of running, breaking every JSON-parsing assertion. cli-beat-writes.test.ts
  already covers the identical cases under the `beat` noun.
- Rename the last three `segment` invocations in
  cli-backup-coordination.test.ts to `beat` for the same reason.
- Harden useLocalStorage's SSR guard: `typeof localStorage !== "undefined"`
  is no longer sufficient — Node now ships a global `localStorage` stub
  whose methods throw "is not a function" unless the process was started
  with `--localstorage-file`. Also check that getItem/setItem exist.
- Fix a real concurrency bug in syncFrozenCourseVersionToDropbox: the
  per-Video upload pool used Effect.forEach's default fail-fast behavior,
  so one Video's rejected upload interrupted sibling Videos still
  mid-transfer, silently dropping their upload-video-complete events.
  Each shipVideo call is now wrapped in Effect.either so every Video gets
  its full turn before the first failure is re-raised. Split the per-Video
  ship pipeline into course-publish-ship-video.ts along the way — it's
  what pushed course-publish-dropbox.ts over the file-token budget.
