+++ app/services/course-publish-dropbox.ts
@@ -1,4 +1,4 @@
-import { Config, Effect, Stream } from "effect";
+import { Config, Effect, Either } from "effect";
 import { FileSystem } from "@effect/platform";
 import { createHash } from "node:crypto";
 import path from "node:path";
@@ -14,13 +14,9 @@ import {
 } from "./export-hash";
 import { VersionOperationsService } from "./db-version-operations.server";
 import { ExportError, PublishValidationError } from "./course-publish-errors";
-import {
-  extractErrorMessage,
-  type EmitPublishDetailEvent,
-} from "./course-publish-export-events";
+import type { EmitPublishDetailEvent } from "./course-publish-export-events";
 import {
   uploadFile,
-  uploadFileFromDisk,
   getMetadata,
   listFolder,
   copyBatch,
@@ -31,50 +27,9 @@ import {
   type ReusePlan,
   type ReusableSource,
 } from "./course-publish-reuse-plan";
-import { DropboxContentHasher } from "./dropbox-content-hash";
 import { getValidDropboxAccessToken } from "./dropbox-auth-service";
 import { uploadConcurrency } from "./dropbox-upload-config";
-import { readExportDigest, writeExportDigest } from "./export-sha256-sidecar";
-
-/**
- * One Video's place in the bundle, all of it read off the DATABASE: where it
- * lands inside the bundle, which local file it comes from, and the Export Hash
- * that addresses that file. Knowable before a single frame has been encoded.
- */
-type VideoEntry = {
-  videoId: string;
-  videoTitle: string;
-  lessonPath: string;
-  localPath: string;
-  relativeAssetPath: string;
-  exportHash: string | null;
-};
-
-/**
- * Read an Exported Video off disk purely to digest it. Only Videos this
- * Publish is NOT sending go through here — anything actually uploaded is
- * digested off the upload's own byte stream instead, so no file is ever read
- * twice and the whole-course pre-hash pass no longer exists.
- */
-const hashFileLocally = Effect.fn("hashFileLocally")(function* (
-  effectFs: FileSystem.FileSystem,
-  filePath: string
-) {
-  const sha256Hash = createHash("sha256");
-  const contentHasher = new DropboxContentHasher();
-  const bytes = yield* effectFs.stream(filePath).pipe(
-    Stream.runFold(0, (total, chunk) => {
-      sha256Hash.update(chunk);
-      contentHasher.update(chunk);
-      return total + chunk.byteLength;
-    })
-  );
-  return {
-    sha256: sha256Hash.digest("hex"),
-    bytes,
-    contentHash: contentHasher.digest(),
-  };
-});
+import { createShipVideo, type VideoEntry } from "./course-publish-ship-video";
 
 /**
  * The handoff for a sync with no export phase in front of it — the manual
@@ -364,224 +319,24 @@ export const syncFrozenCourseVersionToDropbox = Effect.fn(
     lessonPath: string;
   }> = [];
 
-  const emitVideoError = (videoId: string, message: string) =>
-    input.onDetailEvent?.({
-      event: "upload-video-error",
-      data: { videoId, message },
-    });
-
-  /**
-   * Adopt a landed Video with NO local file to check it against.
-   *
-   * This is the case that used to have no answer: an earlier attempt put the
-   * Video at its address, and the export it was made from has since been
-   * collected — so the manifest's SHA256 could not be recovered and the whole
-   * Publish was discarded. Reuse makes that common rather than rare, because
-   * the plan that put the Video there is the same plan that cancelled its
-   * re-encode.
-   *
-   * The previous Bundle answers it. Its manifest owes the new one this Video's
-   * SHA256 and byte count; its listing gives the content hash the landed file
-   * must carry. One comparison, and not a byte read from disk or wire.
-   *
-   * Weaker than `adoptLandedVideo`, and deliberately second to it: where the
-   * source and destination are the same file — an unchanged re-Publish, which
-   * lands in the same Bundle — the comparison is of a file with itself and
-   * cannot detect tampering. Only local bytes can do that, so wherever they
-   * exist they are used instead.
-   */
-  const adoptFromPlan = Effect.fn("adoptVideoFromReusePlan")(function* (
-    entry: VideoEntry,
-    remoteFile: DropboxFileMetadata,
-    source: ReusableSource
-  ) {
-    // Same Export Hash means identical bytes by construction, so the two
-    // hashes must agree. Disagreement is an immutability violation, exactly
-    // as it is for a locally-checked adoption, and is never overwritten.
-    if (
-      remoteFile.content_hash !== source.contentHash ||
-      remoteFile.size !== source.bytes
-    ) {
-      return yield* new ExportError({
-        message: `Immutable asset bundle conflict for video ${entry.videoId}`,
-      });
-    }
-    videoByteSizes.set(entry.videoId, source.bytes);
-    uploadedByVideo.set(entry.videoId, source.bytes);
-    reportProgress();
-    return { sha256: source.sha256, bytes: source.bytes };
+  // One Video's whole trip — wait for its export, then adopt or stream it —
+  // lives in its own module purely to keep this one under the repo's
+  // file-token budget. See course-publish-ship-video.ts for the adoption and
+  // streaming strategies and why a failure there is attributed to its own
+  // Video rather than to the Publish as a whole.
+  const shipVideo = createShipVideo({
+    effectFs,
+    accessToken,
+    onDetailEvent: input.onDetailEvent,
+    cancelledExports: input.cancelledExports,
+    remoteFilesByPath,
+    remoteVideoPath,
+    plannedSourceOf,
+    videoByteSizes,
+    uploadedByVideo,
+    reportProgress,
+    missingVideos,
   });
-
-  /**
-   * A Video already sitting at its address, put there by a previous attempt.
-   * Its SHA256 is still owed to the manifest and nothing streamed it, so the
-   * numbers have to come from somewhere other than the upload: the sidecar
-   * written when this export was last digested, or — if that is missing or
-   * disagrees with the file — one local read that then writes the sidecar for
-   * next time. A mismatch against the remote is an immutability violation
-   * rather than an interrupted transfer, and is never overwritten.
-   */
-  const adoptLandedVideo = Effect.fn("adoptLandedVideo")(function* (
-    entry: VideoEntry,
-    remoteFile: DropboxFileMetadata,
-    fileSize: number
-  ) {
-    const cached = yield* readExportDigest(effectFs, entry.localPath, fileSize);
-    const hashes =
-      cached ?? (yield* hashFileLocally(effectFs, entry.localPath));
-    if (!cached) {
-      yield* writeExportDigest(effectFs, entry.localPath, hashes);
-    }
-    if (
-      remoteFile.content_hash !== hashes.contentHash ||
-      remoteFile.size !== hashes.bytes
-    ) {
-      return yield* new ExportError({
-        message: `Immutable asset bundle conflict for video ${entry.videoId}`,
-      });
-    }
-    // A resumed Publish counts it as done rather than reporting itself back
-    // at zero.
-    uploadedByVideo.set(entry.videoId, hashes.bytes);
-    reportProgress();
-    return { sha256: hashes.sha256, bytes: hashes.bytes };
-  });
-
-  /**
-   * Send the Video's bytes, digesting them off the same pass — the manifest's
-   * proven-source-revision guarantee is met without a separate read.
-   */
-  const streamVideo = Effect.fn("streamVideoToDropbox")(function* (
-    entry: VideoEntry,
-    fileSize: number
-  ) {
-    const sha256Hash = createHash("sha256");
-    const contentHasher = new DropboxContentHasher();
-    let streamedBytes = 0;
-
-    const metadata = yield* uploadFileFromDisk({
-      accessToken,
-      path: remoteVideoPath(entry),
-      filePath: entry.localPath,
-      fileSize,
-      onChunk: (chunk) => {
-        sha256Hash.update(chunk);
-        contentHasher.update(chunk);
-        streamedBytes += chunk.byteLength;
-      },
-      onProgress: (uploaded, total) => {
-        uploadedByVideo.set(entry.videoId, uploaded);
-        reportProgress();
-        input.onDetailEvent?.({
-          event: "upload-video-progress",
-          data: {
-            videoId: entry.videoId,
-            uploadedBytes: uploaded,
-            totalBytes: total,
-          },
-        });
-      },
-    });
-
-    const contentHash = contentHasher.digest();
-    if (metadata.content_hash !== contentHash) {
-      return yield* new ExportError({
-        message: `Upload verification failed for video ${entry.videoId}: content_hash mismatch`,
-      });
-    }
-
-    const digest = {
-      sha256: sha256Hash.digest("hex"),
-      contentHash,
-      bytes: streamedBytes,
-    };
-    // These numbers were free this time — they came off the upload stream. Bank
-    // them so a resumed or unchanged re-Publish, which sends nothing and so
-    // streams nothing, does not have to read the file back to recover them.
-    yield* writeExportDigest(effectFs, entry.localPath, digest);
-
-    return { sha256: digest.sha256, bytes: digest.bytes };
-  });
-
-  /** One Video's whole trip: wait for its export, then skip it or send it. */
-  const shipVideo = Effect.fn("shipVideo")(
-    function* (entry: VideoEntry) {
-      // The handoff: this Video's own export, and nothing else's.
-      yield* input.awaitVideoReady(entry.videoId);
-
-      const remoteFile = remoteFilesByPath.get(
-        remoteVideoPath(entry).toLowerCase()
-      );
-      const plannedSource = plannedSourceOf(entry);
-
-      let receipt: { sha256: string; bytes: number };
-      let onDisk = yield* effectFs.exists(entry.localPath);
-
-      // The local file is the STRONGER witness, so it is always preferred
-      // where it exists: it was produced from this Video's Clips, whereas the
-      // plan can only report what Dropbox already holds. Adopting from the
-      // plan is the fallback for the case that used to have no answer at all.
-      if (!onDisk && remoteFile && plannedSource) {
-        receipt = yield* adoptFromPlan(entry, remoteFile, plannedSource);
-      } else {
-        // A Video the plan cancelled the encode for has arrived here with no
-        // file and nothing at its address, which means its copy did not
-        // happen — the source vanished, or the batch would not run. Nothing
-        // else will ever produce these bytes, so "fall back to upload" has to
-        // mean encoding it now. Without this the saving would turn a slow
-        // Publish into a failed one.
-        if (!onDisk && plannedSource && input.cancelledExports) {
-          yield* input.cancelledExports.restore(entry.videoId);
-          onDisk = yield* effectFs.exists(entry.localPath);
-        }
-
-        if (!onDisk) {
-          missingVideos.push({
-            videoId: entry.videoId,
-            videoTitle: entry.videoTitle,
-            lessonPath: entry.lessonPath,
-          });
-          emitVideoError(entry.videoId, "No exported file to upload");
-          return null;
-        }
-
-        const fileSize = Number((yield* effectFs.stat(entry.localPath)).size);
-        videoByteSizes.set(entry.videoId, fileSize);
-        // This Video has a slot and a size: it is uploading, not queued. The
-        // size rides along from the very first event so a consumer can weight
-        // this Video against its siblings before a byte has moved.
-        input.onDetailEvent?.({
-          event: "upload-video-progress",
-          data: {
-            videoId: entry.videoId,
-            uploadedBytes: 0,
-            totalBytes: fileSize,
-          },
-        });
-
-        receipt = remoteFile
-          ? yield* adoptLandedVideo(entry, remoteFile, fileSize)
-          : yield* streamVideo(entry, fileSize);
-      }
-
-      input.onDetailEvent?.({
-        event: "upload-video-complete",
-        data: { videoId: entry.videoId, bytes: receipt.bytes },
-      });
-      return { videoId: entry.videoId, ...receipt };
-    },
-    // Whatever went wrong for this Video — a missing file, a rejected request,
-    // a hash that did not verify — belongs to that Video by name rather than
-    // to the Publish as an undifferentiated whole.
-    (effect, entry) =>
-      Effect.tapError(effect, (error) => {
-        emitVideoError(
-          entry.videoId,
-          extractErrorMessage(error, "Upload failed unexpectedly")
-        );
-        return Effect.void;
-      })
-  );
 
   /** Receipts owed to the manifest by Videos that were copied, not sent. */
   const copyReceipts = new Map<string, { sha256: string; bytes: number }>();
@@ -655,13 +410,29 @@ export const syncFrozenCourseVersionToDropbox = Effect.fn(
   // excluded from, because it is about to move real bytes after all.
   for (const entry of fallbackEntries) uploadedByVideo.set(entry.videoId, 0);
 
-  const receipts = yield* Effect.forEach(
+  // Each Video gets the whole trip, win or lose: shipVideo is wrapped in
+  // Either so one Video's rejected upload fails ITS slot without Effect.forEach's
+  // default fail-fast interrupting siblings still mid-transfer. Losing that
+  // would mean a Video that finished successfully never gets to report its
+  // own `upload-video-complete` — exactly the "belongs to that Video by name"
+  // guarantee shipVideo's own error handler documents. The first failure is
+  // re-raised below, once every Video has had its turn, so the Publish still
+  // fails overall.
+  const shipResults = yield* Effect.forEach(
     [...uploadingEntries, ...fallbackEntries],
-    shipVideo,
+    (entry) => Effect.either(shipVideo(entry, input.awaitVideoReady)),
     { concurrency: uploadConcurrencyLimit }
   );
+  const receipts = shipResults
+    .filter(Either.isRight)
+    .map((result) => result.right);
+  const shipFailure = shipResults.find(Either.isLeft);
 
   if (missingVideos.length > 0) return { missingVideos };
+
+  if (shipFailure) {
+    return yield* Effect.fail(shipFailure.left);
+  }
 
   const videoAssets = new Map([
     ...copyReceipts,
